refact: renames OpenAPI schema to 'admin-ticket'
