+++ api/src/paths/administrative/tickets/index.ts
@@ -4,7 +4,7 @@ import { SYSTEM_ROLE } from '../../../constants/roles';
 import { getDBConnection } from '../../../database/db';
 import { defaultErrorResponses } from '../../../openapi/schemas/http-responses';
 import { paginationRequestQueryParamSchema } from '../../../openapi/schemas/pagination';
-import { CreateTicketRequestSchema, TicketListResponseSchema, TicketSchema } from '../../../openapi/schemas/ticket';
+import { CreateTicketRequestSchema, TicketListResponseSchema, TicketSchema } from '../../../openapi/schemas/admin-ticket';
 import { authorizeRequestHandler } from '../../../request-handlers/security/authorization';
 import { TicketService } from '../../../services/ticket-service';
 import { getLogger } from '../../../utils/logger';
