+++ api/src/paths/administrative/tickets/{ticketId}/status/index.ts
@@ -3,7 +3,7 @@ import { Operation } from 'express-openapi';
 import { SYSTEM_ROLE } from '../../../../../constants/roles';
 import { getDBConnection } from '../../../../../database/db';
 import { defaultErrorResponses } from '../../../../../openapi/schemas/http-responses';
-import { TicketSchema, UpdateTicketStatusRequestSchema } from '../../../../../openapi/schemas/ticket';
+import { TicketSchema, UpdateTicketStatusRequestSchema } from '../../../../../openapi/schemas/admin-ticket';
 import { authorizeRequestHandler } from '../../../../../request-handlers/security/authorization';
 import { TicketService } from '../../../../../services/ticket-service';
 import { getLogger } from '../../../../../utils/logger';
